create projects.py
