+++ match.py
@@ -3,18 +3,27 @@
 import argparse
 import logging
 from pathlib import Path
-import time
 import json
 
 from pymongo import MongoClient
 
-from dcctools.anonlink import Project, Results
+from dcctools.anonlink import Results
 from dcctools.config import Configuration
 
 # Delay between run status checks
 SLEEP_TIME = 10.0
 
 log = logging.getLogger(__name__)
+
+
+class MissingResults(Exception):
+    def __init__(self, expected_results, available_results, message=None):
+        message = message if message else self.message(expected_results, available_results)
+        super().__init__(message)
+
+    def message(self, expected_results, available_results):
+        missing_results = expected_results.difference(available_results)
+        return f'Missing results for projects: {", ".join(missing_results)}'
 
 
 def parse_args():
@@ -27,115 +36,35 @@ def parse_args():
     parser.add_argument(
         "--verbose", default=False, action="store_true", help="Show debugging output"
     )
-    parser.add_argument(
-        "--projects_only",
-        default=False,
-        action="store_true",
-        help="Run projects and generate results, but don't perform matching"
-    )
-    parser.add_argument(
-        "--match_only",
-        default=False,
-        action="store_true",
-        help="Perform matching based on stored results"
-    )
-    args = parser.parse_args()
-    return args
+    return parser.parse_args()
 
 
-def run_projects(c):
-    if c.household_match:
-        log.debug("Processing households")
-        with open(Path(c.household_schema)) as schema_file:
-            household_schema = schema_file.read()
-            project_name = "fn-phone-addr-zip"
-            household_project = Project(
-                project_name,
-                household_schema,
-                c.systems,
-                c.entity_service_url,
-                c.blocked,
-            )
-            household_project.start_project()
-            for system in c.systems:
-                household_project.upload_clks(
-                    system, c.get_household_clks_raw(system, project_name)
-                )
-            if type(c.matching_threshold) == list:
-                threshold = c.matching_threshold[0]
-            else:
-                threshold = c.matching_threshold
-            household_project.start_run(threshold)
-            running = True
-            print("\n--- RUNNING ---\n")
-            while running:
-                status = household_project.get_run_status()
-                print(status)
-                if status.get("state") == "completed":
-                    running = False
-                    break
-                time.sleep(SLEEP_TIME)
-            print("\n--- Getting results ---\n")
-            result_json = household_project.get_results()
-            Path(c.project_results_dir).mkdir(parents=True, exist_ok=True)
-            with open(Path(c.project_results_dir) / f'{project_name}.json', 'w') as json_file:
-                json.dump(result_json, json_file)
+def has_results_available(config, projects=None):
+    available_results = set(map(lambda x: x.stem, Path(config.project_results_dir).glob('*.json')))
+    expected_results = set(projects) if projects else set(config.projects)
+    if (expected_results <= available_results):
+        return True
     else:
-        log.debug("Processing individuals")
-        if c.blocked:
-            log.debug("Blocked, extracting CLKs and blocks")
-            for system in c.systems:
-                c.extract_clks(system)
-                c.extract_blocks(system)
-
-        iter_num = 0
-        for i, (project_name, schema) in enumerate(c.load_schema().items()):
-            iter_num = iter_num + 1
-            project = Project(
-                project_name, schema, c.systems, c.entity_service_url, c.blocked
-            )
-            project.start_project()
-            for system in c.systems:
-                if c.blocked:
-                    project.upload_clks_blocked(
-                        system,
-                        c.get_clk(system, project_name),
-                        c.get_block(system, project_name),
-                    )
-                else:
-                    project.upload_clks(system, c.get_clks_raw(system, project_name))
-            if type(c.matching_threshold) == list:
-                threshold = c.matching_threshold[i]
-            else:
-                threshold = c.matching_threshold
-            project.start_run(threshold)
-            running = True
-            print("\n--- RUNNING ---\n")
-            while running:
-                status = project.get_run_status()
-                print(status)
-                if status.get("state") == "completed":
-                    running = False
-                    break
-                time.sleep(SLEEP_TIME)
-            print("\n--- Getting results ---\n")
-            result_json = project.get_results()
-            Path(c.project_results_dir).mkdir(parents=True, exist_ok=True)
-            with open(Path(c.project_results_dir) / f'{project_name}.json', 'w') as json_file:
-                json.dump(result_json, json_file)
+        raise MissingResults(expected_results, available_results)
 
 
-def do_match(systems, project_results_dir, database, household_match):
-    for file_name in Path(project_results_dir).glob('*.json'):
-        project_name = Path(file_name).stem
-        with open(file_name) as file:
+def do_match(config, projects=None):
+    projects = projects if projects else config.projects
+    database = MongoClient(config.mongo_uri).linkage_agent
+    if config.household_match:
+        do_matching(config, ["fn-phone-addr-zip"], database.household_match_groups)
+    else:
+        do_matching(config, projects, database.match_groups)
+
+
+def do_matching(config, projects, collection):
+    has_results_available(config, projects)
+    for project_name in projects:
+        with open(Path(config.project_results_dir) / f'{project_name}.json') as file:
             result_json = json.load(file)
-            results = Results(systems, project_name, result_json)
-            print('Matching groups for system f{project_name}')
-            if household_match:
-                results.insert_results(database.match_groups)
-            else:
-                results.insert_results(database.house_hold_match_groups)
+            results = Results(config.systems, project_name, result_json)
+            print(f'Matching for project: {project_name}')
+            results.insert_results(collection)
 
 
 if __name__ == "__main__":
@@ -146,11 +75,4 @@ if __name__ == "__main__":
 
     client = MongoClient(config.mongo_uri)
     database = client.linkage_agent
-
-    if args.projects_only:
-        run_projects(config)
-    elif args.match_only:
-        do_match(config.systems, config.project_results_dir, database, config.household_match)
-    else:
-        run_projects(config)
-        do_match(config.systems, config.project_results_dir, database, config.household_match)
+    do_match(config)
