+++ test_util/linkage/run_full_linkage_test.py
@@ -9,6 +9,8 @@ def run_full_linkage_test(local_config_file):
     lau.drop(config)
     print("Doing validation...")
     lau.validate(config)
+    print("Running projects...")
+    lau.projects(config)
     print("Doing match...")
     lau.match(config)
     print("Doing link_id...")
